+++ tests/core/test_path_package_functions.py
@@ -4,9 +4,8 @@ from pytest import fixture, mark, raises
 
 from seedcase_sprout.core import (
     path_package,
-    path_package_database,
-    path_package_properties,
     path_packages,
+    path_properties,
 )
 from tests.core.directory_structure_setup import (
     create_test_package_structure,
@@ -27,8 +26,7 @@ def tmp_sprout_global(monkeypatch, tmp_path):
     "function, expected_path",
     [
         (path_package, "packages/1"),
-        (path_package_database, "packages/1/database.sql"),
-        (path_package_properties, "packages/1/datapackage.json"),
+        (path_properties, "packages/1/datapackage.json"),
     ],
 )
 def test_path_package_functions_return_expected_path(
@@ -41,7 +39,7 @@ def test_path_package_functions_return_expected_path(
 
 @mark.parametrize(
     "function",
-    [path_package, path_package_database, path_package_properties],
+    [path_package, path_properties],
 )
 def test_path_package_functions_raise_error_if_package_id_does_not_exist(
     tmp_sprout_global, function